Criação dos endpoints para converterArquivos e persistir dados no banco

+++ src/main/kotlin/com/br/labs/model/Produto.kt
@@ -12,5 +12,5 @@ data class Produto(
     val idProduto: Int? = null,
 
     @Column(name = "valor", length = 12)
-    val valor: BigDecimal
+    val valor: Double
 )
